# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HNG Chrome extension - Screen Recording API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL"
        }
    ],
    "paths": {
        "/api/v1/recordings": {
            "get": {
                "tags": [
                    "Recordings"
                ],
                "summary": "List all recordings",
                "description": "Returns a list of all recordings.",
                "operationId": "e96554e8c32de40677656ddaf5b75c96",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RecordingResource"
                                    }
                                },
                                "example": [
                                    {
                                        "id": "9a3e8d15-b805-4309-989b-bba4f78a9248",
                                        "title": "Kizz Daniel - Jaho (Official Video)",
                                        "url": "http://localhost:8000/storage/videos/1695934456_kizz_daniel_-_jaho_(official_video).mp4",
                                        "description": null,
                                        "fileName": "Kizz Daniel - Jaho (Official Video).mp4",
                                        "fileSize": "10690611",
                                        "thumbnail": null,
                                        "slug": "kizz-daniel-jaho-official-video",
                                        "createdAt": "2023-09-28T20:54:16.000000Z"
                                    },
                                    {
                                        "id": "9a3e72a0-70d6-4350-a6b3-4955dac13db2",
                                        "title": "New Recording",
                                        "url": "http://localhost:8000/storage/videos/1695930017_kizz_daniel_-_jaho_(official_video).mp4",
                                        "description": null,
                                        "fileName": "Kizz Daniel - Jaho (Official Video).mp4",
                                        "fileSize": "10690611",
                                        "thumbnail": "http://localhost:8000/storage/thumbnails/1695930017_yos3.png",
                                        "slug": "new-recording",
                                        "createdAt": "2023-09-28T19:40:17.000000Z"
                                    },
                                    {
                                        "id": "9a3e6efe-9fd6-4570-9781-d99b1e8eb446",
                                        "title": "Kizz Daniel - Jaho (Official Video)",
                                        "url": "http://localhost:8000/storage/videos/1695929408_kizz_daniel_-_jaho_(official_video).mp4",
                                        "description": null,
                                        "fileName": "Kizz Daniel - Jaho (Official Video).mp4",
                                        "fileSize": "10690611",
                                        "thumbnail": "http://localhost:8000/storage/thumbnails/1695929408_yos3.png",
                                        "slug": "kizz-daniel-jaho-official-video",
                                        "createdAt": "2023-09-28T19:30:08.000000Z"
                                    }
                                ]
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops something went wrong"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recordings": {
            "post": {
                "tags": [
                    "Recordings"
                ],
                "summary": "Create a new recording",
                "description": "Creates a new recording.",
                "operationId": "f8e75c34d07362a67c93ca92ba78b70b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recording created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recording created successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RecordingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error message"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops something went wrong"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recordings/{recording}": {
            "get": {
                "tags": [
                    "Recordings"
                ],
                "summary": "Get a recording by ID",
                "description": "Returns details of a recording by its ID.",
                "operationId": "895af8c76d5ac9ef76c2ae883c961ecc",
                "parameters": [
                    {
                        "name": "recording",
                        "in": "path",
                        "description": "Recording ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecordingResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recording not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Recording not found"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops something went wrong"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/recordings/{id}": {
            "delete": {
                "tags": [
                    "Recordings"
                ],
                "summary": "Delete a recording by ID",
                "description": "Deletes a recording by its ID.",
                "operationId": "92f390d7d84012b66f8c062410529ca9",
                "parameters": [
                    {
                        "name": "recording",
                        "in": "path",
                        "description": "Recording ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recording deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recording deleted successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recording not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Recording not found"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops something went wrong"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RecordingResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9a3e8d15-b805-4309-989b-bba4f78a9248"
                    },
                    "title": {
                        "type": "string",
                        "example": "Kizz Daniel - Jaho (Official Video)"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost:8000/storage/videos/1695934456_kizz_daniel_-_jaho_(official_video).mp4"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "fileName": {
                        "type": "string",
                        "example": "Kizz Daniel - Jaho (Official Video).mp4"
                    },
                    "fileSize": {
                        "type": "string",
                        "example": "10690611"
                    },
                    "thumbnail": {
                        "type": "string",
                        "format": "uri",
                        "example": null,
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "kizz-daniel-jaho-official-video"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-28T20:54:16.000000Z"
                    }
                },
                "type": "object"
            },
            "RecordingRequest": {
                "description": "Recording request data",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Sample Title",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Description of the recording",
                        "nullable": true
                    },
                    "file": {
                        "description": "The video file to upload",
                        "type": "string",
                        "format": "binary",
                        "nullable": false
                    },
                    "thumbnail": {
                        "description": "[Optional] thumbnail image file to upload",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    }
}